Update the reservation lists in redux
